+++ src/components/CollectionItem.vue
@@ -1,0 +1,71 @@
+<template>
+  <div class="card provider-card cell small">
+    <div class="card-divider">
+      <span class="provider-name">{{ provider.display_name }}</span>
+    </div>
+    <div class="provider-logo">
+      <a :href="'/collections/'+provider.provider_name">
+        <img :alt="provider"
+            :src="getProviderLogo(provider.provider_name)">
+      </a>
+    </div>
+    <div class="card-section">
+      <span>Collection size: {{ getProviderImageCount(provider.image_count) }} images</span>
+    </div>
+  </div>
+</template>
+
+<script>
+import ImageProviderService from '@/api/ImageProviderService';
+
+export default {
+  name: 'collection-item',
+  props: ['provider'],
+  methods: {
+    getProviderImageCount(imageCount) {
+      return (imageCount).toLocaleString('en');
+    },
+    getProviderLogo(providerName) {
+      const provider = ImageProviderService.getProviderInfo(providerName);
+      if (provider) {
+        const logo = provider.logo;
+        const logoUrl = require(`@/assets/${logo}`); // eslint-disable-line global-require, import/no-dynamic-require
+
+        return logoUrl;
+      }
+
+      return '';
+    },
+  },
+};
+</script>
+
+<style lang="scss" scoped>
+  .provider-card {
+    width: 18em;
+    background-color: #dedede;
+    margin: 0.5em;
+  }
+
+  .provider-name {
+    font-weight: 800;
+  }
+
+  .provider-logo {
+    height: 10em;
+    line-height: 10em;
+    white-space: nowrap;
+    position: relative;
+
+    a {
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+    }
+
+    img {
+      width: 100%;
+    }
+  }
+</style>
